+++ ChangeLog
@@ -1,11 +1,12 @@
 2016-06-17  Daniel Standage <daniel.standage@gmail.com>
 
    * scripts/filter-abund-single.py: add -o/--outfile option.
-   * tests/test_script_output.py: mov function to calculate file MD5 hashes to
-   test utils module (tests/khmer_tst_utils.py).
-   * tests/test_filter_abund.py: add minimal test for new -o option.
+   * tests/test_script_output.py: move `_calc_md5` function for calculating file
+   MD5 hashes to test utils module (tests/khmer_tst_utils.py).
+   * tests/{test_filter_abund.py,test_scripts.py): move filter_abund tests to
+   dedicated test script, add minimal test for new -o option.
    * tests/test-data/paired-mixed-witherror.fa.pe: add data file in support of
-   new test
+   new test.
    * .gitignore: add .cache/ directory, which appears to be an artifact of the
    py.test framework.
